+++ uui-editor/src/index.ts
@@ -3,6 +3,6 @@ export * from './defaultPlugins';
 export * from './plugins';
 export * from './types';
 export * from './helpers';
-export * from './migrations/slate_migrations';
+export * from './migrations';
 export * from './serialization';
 export * from './implementation/ToolbarButton';