+++ uui-editor/src/migrations/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './slate_migrations';
+export * from './plate_migrations';